// libraries/schema/src/postgres/identity/auth-oauth.generated.ts
export const oauthClients = pgTable("oauth_client", {
		id: uuid().default(sql`uuidv7()`).primaryKey(),
		clientId: text().notNull(),
		clientSecret: text(),
		clientDiscoveryId: text(),
		disabled: boolean(),
		skipConsent: boolean(),
		enableEndSession: boolean(),
		subjectType: text(),
		scopes: text().array(),
		clientCredentialsScopes: text().array(),
		userId: uuid(),
		createdAt: timestamp({ withTimezone: true, precision: 3, mode: "date" }),
		updatedAt: timestamp({ withTimezone: true, precision: 3, mode: "date" }),
		name: text(),
		uri: text(),
		icon: text(),
		contacts: text().array(),
		tos: text(),
		policy: text(),
		softwareId: text(),
		softwareVersion: text(),
		softwareStatement: text(),
		redirectUris: text().array().notNull(),
		postLogoutRedirectUris: text().array(),
		backchannelLogoutUri: text(),
		backchannelLogoutSessionRequired: boolean(),
		tokenEndpointAuthMethod: text(),
		applicationType: text(),
		jwks: text(),
		jwksUri: text(),
		grantTypes: text().array(),
		responseTypes: text().array(),
		requirePKCE: boolean(),
		dpopBoundAccessTokens: boolean(),
		referenceId: text(),
		metadata: jsonb().$type<unknown>(),
	}, (table): PgTableExtraConfigValue[] => [
		foreignKey({ name: "oauth_client_authority_fk", columns: [table.id], foreignColumns: [oauthClientAuthority.id] }).onDelete("restrict"),
		uniqueIndex("oauth_client_client_id_key").on(table.clientId),
		index("oauth_client_user_id_idx").on(table.userId),
		foreignKey({ name: "oauth_client_user_id_user_id_fk", columns: [table.userId], foreignColumns: [users.id] }),
	],
);